Add test regress and rearrange test names
