+++ tests/test_02_01_resize_ls_oblique_java.py
