+++ tests/test_02_02_resize_ls_oblique.py
