+++ tests/test_05_01_smooth.py
